use pgoapi from requirements.txt instead of copying its files
