--- pogom/pgoapi/__init__.py
@@ -1,5 +1,0 @@
-from pgoapi      import *
-
-
-import requests.packages.urllib3
-requests.packages.urllib3.disable_warnings()
