+++ pogom/utils.py
@@ -11,6 +11,10 @@ import logging
 import shutil
 import requests
 
+from importlib import import_module
+from s2sphere import LatLng, CellId
+from geopy.geocoders import GoogleV3
+
 from . import config
 
 log = logging.getLogger(__name__)
@@ -272,3 +276,36 @@ def send_to_webhook(message_type, message):
                 log.debug('Response timeout on webhook endpoint %s', w)
             except requests.exceptions.RequestException as e:
                 log.debug(e)
+
+def get_pos_by_name(location_name):
+    prog = re.compile("^(\-?\d+\.\d+)?,\s*(\-?\d+\.\d+?)$")
+    res = prog.match(location_name)
+    latitude, longitude, altitude = None, None, None
+    if res:
+        latitude, longitude, altitude = float(res.group(1)), float(res.group(2)), 0
+    elif location_name:
+        geolocator = GoogleV3()
+        loc = geolocator.geocode(location_name)
+        if loc:
+            latitude, longitude, altitude = loc.latitude, loc.longitude, loc.altitude
+
+    return (latitude, longitude, altitude)
+
+def get_class(cls):
+    module_, class_ = cls.rsplit('.', 1)
+    class_ = getattr(import_module(module_), class_)
+    return class_
+
+def get_cell_ids(lat, long):
+    origin = CellId.from_lat_lng(LatLng.from_degrees(lat, long)).parent(15)
+    walk = [origin.id()]
+
+    # 10 before and 10 after
+    next = origin.next()
+    prev = origin.prev()
+    for i in range(10):
+        walk.append(prev.id())
+        walk.append(next.id())
+        next = next.next()
+        prev = prev.prev()
+    return sorted(walk)
