+++ runserver.py
@@ -35,7 +35,7 @@ from pogom.utils import get_args, insert_mock_data
 from pogom.search import search_overseer_thread, fake_search_loop
 from pogom.models import init_database, create_tables, drop_tables, Pokemon, Pokestop, Gym
 
-from pogom.pgoapi.utilities import get_pos_by_name
+from pogom.utils import get_pos_by_name
 
 
 if __name__ == '__main__':
